Moved utils.GetTopologyJSON and related structs to server_utils package
